fix: Fix nullable reference warnings in Table_JSTF.cs and Table_hmtx.cs
- Table_JSTF.cs: Fixed 11 methods to use nullable return types
  - GetExtenderGlyphTable, GetDefJstfLangSysTable, GetJstfLangSysRecord
  - GetJstfLangSysTable, GetShrinkageEnableGSUBTable, GetShrinkageDisableGSUBTable
  - GetShrinkageEnableGPOSTable, GetShrinkageDisableGPOSTable, GetShrinkageJstfMaxTable
  - GetExtensionEnableGSUBTable, GetExtensionDisableGSUBTable, GetExtensionEnableGPOSTable
  - GetExtensionDisableGPOSTable, GetExtensionJstfMaxTable, GetJstfScriptRecord
- Table_hmtx.cs: Fixed nullable warnings (from previous session)
- Eliminated all CS86 warnings from these files

+++ OTFontFile/src/Table_hmtx.cs
@@ -167,7 +167,7 @@ namespace OTFontFile
                     for (uint i=(uint)nFirstWidthPos+1; i<numberOfHMetrics; i++)
                     {
                         longHorMetric? hm = GetHMetric(i, fontOwner);
-                        ushort nWidth = hm != null ? hm.advanceWidth : 0;
+                        ushort nWidth = hm != null ? hm.advanceWidth : (ushort)0;
                         if (nWidth != 0 && nWidth != nFirstWidth)
                         {
                             bHmtxMono = false;
